Compute dissimilarity estimates between two samples based on Expected Species Shared (ESS)-measures, using abundance data for the species contained in each samples

Compute dissimilarity estimates between two samples based on Expected Species Shared (ESS)-measures, using abundance data for the species contained in each samples

Usage

ess(x, m = 1, index = "CNESSa")

Arguments

x

a community data matrix (sample x species); sample name is the row name of the matrix

m

the sample size parameter that represents the number of individuals randomly drawn from each sample, which by default is set to m=1, but can be changed according to the users' requirements. Rows with a total sample size <m will be excluded automatically from the analysis.

index

the distance measure used in the calculation, as one of the four options "CNESSa", "CNESS","NESS" and "ESS", with the default set as "CNESSa"

Value

a pair-wised matrix

References

Zou Y, Axmacher JC (2020). “The Chord-Normalized Expected Species Shared (CNESS)-distance represents a superior measure of species turnover patterns.” Methods in Ecology and Evolution, 11(2), 273-280. doi:10.1111/2041-210X.13333 .

Examples

data(share, package = 'rarestR')
ess(share)
#>           1         2
#> 2 0.7970962          
#> 3 0.6359703 0.7642330
ess(share, m = 100)
#>           1         2
#> 2 0.8566624          
#> 3 0.7308390 0.8229221
ess(share, m = 100, index = "ESS")
#>          1        2
#> 2 13.01735         
#> 3 22.65674 13.23924