Calculation of Total Expected Species base on ESa, ESb and their average value

Calculation of Total Expected Species base on ESa, ESb and their average value

Usage

tes(x)

Arguments

x

a data vector representing number of individuals for each species

Value

a list in a self-defined class 'rarestr'. See "Details".

Details

The value returned by the tes() function in the 'rarestr' class is a list containing three parts:

par

A data frame of the summary of the estimated values and their standard deviations based on TESa, TESb, and TESab, and the model used in the estimation of TES, either 'logistic' or 'Weibull'.

TESa

A list of the modeled results with the TESa method.

TESb

A list of the modeled results with the TESb method.

Both TESa and TESb contain five parts, including a data frame of the parameters ($par), a data frame of the simulated results ($result), a maximum x value ($xmax), a vector of the predicted x value ($Predx), and a vector of the predicted y value ($Predy)

References

zou2023rarestR

Examples

data(share, package = 'rarestR')
Output_tes <- tes(share[1,])
Output_tes
#>          est est.sd model.par
#> TESa  138.50   2.46  logistic
#> TESb   92.63  32.65   Weibull
#> TESab 115.56  16.37      <NA>