Calculate the Total number of Expected Shared Species between two samples.

Calculate the Total number of Expected Shared Species between two samples.

tess(x, knots = 40, plot = FALSE)

Arguments

x

a data matrix for two samples representing two communities (plot x species)

knots

specifies the number of separate sample sizes of increasing value used for the calculation of ESS between 1 and the endpoint, which by default is set to knots=40

plot

a logical value that provides the fitted curve

method

the method the model used, with three options available as "auto", "logistic" and "weibull", with the default set as "auto"

Value

The asymptotic value (a), the model used in the estimation of TESS (either "Three" or "Four") and the model fit (r.sq).