Calculate the Total number of Expected Shared Species between two samples.

Calculate the Total number of Expected Shared Species between two samples.

Usage

tess(x)

Arguments

x

a data matrix for two samples representing two communities (plot x species)

Value

a list in a self-defined class 'rarestr'. See "Details".

Details

The value returned by the tess() function in the 'rarestr' class is a list containing five parts:

par

A data frame of the summary of the estimated values and their standard deviations based on TESa, TESb, and TESab, and the model used in the estimation of TES, either 'logistic' or 'Weibull'.

result

A data frame of the simulated results.

xmax

A maximum x value.

Predx

A vector of the predicted x value.

Predy

A vector of the predicted y value.

References

zou2021rarestR

Examples

data(share, package = 'rarestR')
Output_tess <- tess(share[1:2,])
Output_tess
#>     est est.sd model.par
#> 1 23.28   2.59  logistic