Calculate the Total number of Expected Shared Species between two samples.

Calculate the Total number of Expected Shared Species between two samples.

Usage

tess(x, knots = 40)

Arguments

x

a data matrix for two samples representing two communities (plot x species)

knots

specifies the number of separate sample sizes of increasing value used for the calculation of ESS between 1 and the endpoint, which by default is set to knots=40

Value

estimated values and their standard deviations of TESS, and the model used in the estimation of TES, either 'logistic' or 'Weibull'

References

Zou Y, Axmacher JC (2021). “Estimating the number of species shared by incompletely sampled communities.” Ecography, 44(7), 1098-1108. doi:10.1111/ecog.05625 .

Examples

data(share, package = 'rarestR')
Output_tess <- tess(share[1:2,])
Output_tess
#>     est est.sd model.par
#> 1 23.28   2.59  logistic